refactor: split long source files into sub-modules
- Split src/tools/bash.rs (1733→752): extract background task subsystem
  and shell command analysis into bash/background.rs and bash/commands.rs
- Split src/tui/command.rs (1665→926): extract status display rendering
  into command/display.rs and model help into command/model.rs

All external APIs preserved via pub use re-exports.
