+++ src/tools/bash.rs
@@ -1,697 +1,33 @@
 use crate::sandbox::{SandboxManager, WrappedCommand, sandbox_failure_hint};
 use crate::tool::{Tool, ToolError, ToolOutputStream, ToolProgressEvent};
 use async_trait::async_trait;
-use serde::{Deserialize, Serialize};
 use serde_json::{Value, json};
 use std::collections::HashMap;
-use std::env;
-use std::io::SeekFrom;
 use std::path::Path;
-use std::path::PathBuf;
 use std::process::Stdio;
 use std::sync::Arc;
-use std::sync::Mutex;
 use tokio::fs;
-use tokio::io::{AsyncRead, AsyncReadExt, AsyncSeekExt, AsyncWriteExt};
-use tokio::process::{Child, Command};
+use tokio::process::Command;
 use tokio::sync::mpsc;
-use tokio::sync::oneshot;
-use uuid::Uuid;
+
+mod commands;
+mod background;
+
+pub use self::background::{
+    BackgroundTaskListTool, BackgroundTaskStatusTool, BackgroundTaskStatus,
+    BackgroundTaskStopTool, BackgroundTaskStore, BackgroundTaskRecord,
+    read_output_tail, spawn_background_bash_task, read_stream_chunks,
+};
+pub use self::commands::{
+    BashCommandInput, BashStreamKind,
+    sandbox_command_env, command_env_for_wrapped,
+    sandbox_output_hint, unsandboxed_execution_warning, ensure_sandbox_home_dirs,
+};
 
 pub struct BashTool {
     pub sandbox: Arc<SandboxManager>,
     pub background_tasks: Arc<BackgroundTaskStore>,
     pub base_env: Arc<HashMap<String, String>>,
-}
-
-pub struct BackgroundTaskStatusTool {
-    pub background_tasks: Arc<BackgroundTaskStore>,
-}
-
-pub struct BackgroundTaskListTool {
-    pub background_tasks: Arc<BackgroundTaskStore>,
-}
-
-pub struct BackgroundTaskStopTool {
-    pub background_tasks: Arc<BackgroundTaskStore>,
-}
-
-#[derive(Debug, Clone)]
-pub struct BackgroundTaskStore {
-    dir: PathBuf,
-    tasks: Arc<Mutex<HashMap<String, BackgroundTaskRecord>>>,
-    stop_signals: Arc<Mutex<HashMap<String, oneshot::Sender<()>>>>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct BackgroundTaskRecord {
-    id: String,
-    command: String,
-    output_path: PathBuf,
-    status: BackgroundTaskStatus,
-    exit_code: Option<i32>,
-    sandboxed: bool,
-    sandbox_backend: String,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum BackgroundTaskStatus {
-    Running,
-    Completed,
-    Failed,
-    Killed,
-}
-
-#[derive(Clone, Copy, Debug)]
-enum BashStreamKind {
-    Stdout,
-    Stderr,
-}
-
-impl BashStreamKind {
-    fn output_stream(self) -> ToolOutputStream {
-        match self {
-            Self::Stdout => ToolOutputStream::Stdout,
-            Self::Stderr => ToolOutputStream::Stderr,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct BashCommandInput {
-    #[serde(default)]
-    pub command: Option<String>,
-    #[serde(default)]
-    pub program: Option<String>,
-    #[serde(default)]
-    pub args: Vec<String>,
-    #[serde(default)]
-    pub cwd: Option<String>,
-    #[serde(default)]
-    pub env: HashMap<String, String>,
-    #[serde(default)]
-    pub allow_net: bool,
-    #[serde(default)]
-    pub run_in_background: bool,
-}
-
-impl BashCommandInput {
-    pub fn from_value(input: Value) -> Result<Self, ToolError> {
-        let parsed: Self = serde_json::from_value(input)
-            .map_err(|err| ToolError::InvalidInput(format!("bash payload: {err}")))?;
-        parsed.validate()?;
-        Ok(parsed)
-    }
-
-    pub fn validate(&self) -> Result<(), ToolError> {
-        let has_command = self
-            .command
-            .as_ref()
-            .is_some_and(|value| !value.trim().is_empty());
-        let has_program = self
-            .program
-            .as_ref()
-            .is_some_and(|value| !value.trim().is_empty());
-        if !has_command && !has_program {
-            return Err(ToolError::InvalidInput(
-                "bash payload requires either command or program".into(),
-            ));
-        }
-        Ok(())
-    }
-
-    pub fn working_dir(&self) -> Result<String, ToolError> {
-        match self.cwd.as_ref() {
-            Some(cwd) if !cwd.trim().is_empty() => Ok(cwd.clone()),
-            _ => Ok(env::current_dir()?.to_string_lossy().to_string()),
-        }
-    }
-
-    pub fn summary(&self) -> String {
-        if let Some(command) = self
-            .command
-            .as_ref()
-            .map(|value| value.trim())
-            .filter(|value| !value.is_empty())
-        {
-            return command.to_string();
-        }
-
-        let program = self
-            .program
-            .as_deref()
-            .filter(|value| !value.trim().is_empty())
-            .unwrap_or("<program>");
-        if self.args.is_empty() {
-            program.to_string()
-        } else {
-            format!("{program} {}", self.args.join(" "))
-        }
-    }
-
-    pub fn to_value(&self) -> Value {
-        serde_json::to_value(self).expect("bash command input should serialize")
-    }
-
-    pub fn is_read_only(&self) -> bool {
-        if self.allow_net || self.run_in_background || !self.env.is_empty() {
-            return false;
-        }
-        if let Some(command) = self
-            .command
-            .as_ref()
-            .map(|value| value.trim())
-            .filter(|value| !value.is_empty())
-        {
-            return shell_command_is_read_only(command);
-        }
-        self.program
-            .as_deref()
-            .filter(|value| !value.trim().is_empty())
-            .is_some_and(|program| argv_is_read_only(program, &self.args))
-    }
-
-    pub fn approval_prefix(&self) -> Option<String> {
-        if let Some(command) = self
-            .command
-            .as_ref()
-            .map(|value| value.trim())
-            .filter(|value| !value.is_empty())
-        {
-            let segments = split_shell_segments(command)?;
-            if segments.len() != 1 {
-                return None;
-            }
-            let tokens = tokenize_shell_segment(&segments[0])?;
-            return prefix_from_tokens(&tokens);
-        }
-
-        let program = self
-            .program
-            .as_deref()
-            .filter(|value| !value.trim().is_empty())?;
-        let mut tokens = Vec::with_capacity(self.args.len() + 1);
-        tokens.push(program.to_string());
-        tokens.extend(self.args.iter().cloned());
-        prefix_from_tokens(&tokens)
-    }
-
-    pub fn matches_approval_prefix(&self, prefix: &str) -> bool {
-        let normalized = self.normalized_approval_summary();
-        normalized == prefix
-            || normalized
-                .strip_prefix(prefix)
-                .is_some_and(|suffix| suffix.starts_with(char::is_whitespace))
-    }
-
-    fn normalized_approval_summary(&self) -> String {
-        if let Some(command) = self
-            .command
-            .as_ref()
-            .map(|value| value.trim())
-            .filter(|value| !value.is_empty())
-        {
-            if let Some(tokens) = split_shell_segments(command).and_then(|segments| {
-                if segments.len() == 1 {
-                    tokenize_shell_segment(&segments[0])
-                } else {
-                    None
-                }
-            }) {
-                return normalized_tokens_summary(&tokens);
-            }
-            return command.to_string();
-        }
-
-        let Some(program) = self
-            .program
-            .as_deref()
-            .filter(|value| !value.trim().is_empty())
-        else {
-            return self.summary();
-        };
-        let mut tokens = Vec::with_capacity(self.args.len() + 1);
-        tokens.push(program.to_string());
-        tokens.extend(self.args.iter().cloned());
-        normalized_tokens_summary(&tokens)
-    }
-}
-
-fn prefix_from_tokens(tokens: &[String]) -> Option<String> {
-    let program = tokens.first()?;
-    let program = command_basename(program);
-    if let Some(subcommand) = approval_subcommand_token(program, &tokens[1..]) {
-        Some(format!("{program} {subcommand}"))
-    } else {
-        Some(program.to_string())
-    }
-}
-
-fn normalized_tokens_summary(tokens: &[String]) -> String {
-    let Some(program) = tokens.first() else {
-        return String::new();
-    };
-    let program = command_basename(program);
-    let rest = &tokens[1..];
-    let args = approval_subcommand_index(program, rest)
-        .map(|index| rest[index..].iter().cloned().collect::<Vec<_>>())
-        .unwrap_or_else(|| rest.to_vec());
-    std::iter::once(program.to_string())
-        .chain(args)
-        .collect::<Vec<_>>()
-        .join(" ")
-}
-
-fn command_basename(command: &str) -> &str {
-    Path::new(command)
-        .file_name()
-        .and_then(|name| name.to_str())
-        .unwrap_or(command)
-}
-
-fn approval_subcommand_token<'a>(program: &str, args: &'a [String]) -> Option<&'a str> {
-    approval_subcommand_index(program, args).and_then(|index| args.get(index).map(String::as_str))
-}
-
-fn approval_subcommand_index(program: &str, args: &[String]) -> Option<usize> {
-    match program {
-        "git" => skip_known_global_options(
-            args,
-            &["--no-pager", "--no-optional-locks"],
-            &["-C", "-c", "--git-dir", "--work-tree"],
-        ),
-        "docker" => skip_known_global_options(
-            args,
-            &["--debug", "--tls", "--tlsverify"],
-            &["--config", "--context", "--host", "-H", "--log-level"],
-        ),
-        _ => args.first().map(|_| 0),
-    }
-}
-
-fn skip_known_global_options(
-    args: &[String],
-    valueless_options: &[&str],
-    value_options: &[&str],
-) -> Option<usize> {
-    let mut index = 0;
-    while index < args.len() {
-        let arg = args[index].as_str();
-        if valueless_options.contains(&arg) {
-            index += 1;
-        } else if value_options.contains(&arg) {
-            index += 2;
-        } else if value_options
-            .iter()
-            .any(|option| arg.starts_with(&format!("{option}=")))
-        {
-            index += 1;
-        } else if arg.starts_with('-') {
-            index += 1;
-        } else {
-            return Some(index);
-        }
-    }
-    None
-}
-
-fn shell_command_is_read_only(command: &str) -> bool {
-    if command.contains('\n')
-        || command.contains('`')
-        || command.contains("$(")
-        || command.contains('>')
-    {
-        return false;
-    }
-    split_shell_segments(command)
-        .filter(|segments| !segments.is_empty())
-        .is_some_and(|segments| {
-            segments.into_iter().all(|segment| {
-                tokenize_shell_segment(&segment).is_some_and(|tokens| {
-                    if tokens.is_empty() {
-                        return false;
-                    }
-                    argv_is_read_only(&tokens[0], &tokens[1..])
-                })
-            })
-        })
-}
-
-fn argv_is_read_only(program: &str, args: &[String]) -> bool {
-    let program = Path::new(program)
-        .file_name()
-        .and_then(|name| name.to_str())
-        .unwrap_or(program);
-    match program {
-        "pwd" | "ls" | "tree" | "cat" | "head" | "tail" | "wc" | "stat" | "file" | "du" | "df"
-        | "which" | "type" | "whereis" | "uname" => true,
-        "rg" | "grep" => !args
-            .iter()
-            .any(|arg| matches!(arg.as_str(), "--files-with-matches=")),
-        "sed" => !args.iter().any(|arg| {
-            arg == "-i"
-                || arg.starts_with("-i.")
-                || arg == "--in-place"
-                || arg.starts_with("--in-place=")
-        }),
-        "find" => !args.iter().any(|arg| {
-            matches!(
-                arg.as_str(),
-                "-delete" | "-exec" | "-execdir" | "-ok" | "-okdir"
-            )
-        }),
-        "fd" | "fdfind" => !args.iter().any(|arg| {
-            matches!(
-                arg.as_str(),
-                "-x" | "--exec" | "-X" | "--exec-batch" | "--list-details"
-            )
-        }),
-        "git" => git_args_are_read_only(args),
-        "docker" => docker_args_are_read_only(args),
-        "pyright" => !args
-            .iter()
-            .any(|arg| matches!(arg.as_str(), "--watch" | "-w")),
-        _ => false,
-    }
-}
-
-fn git_args_are_read_only(args: &[String]) -> bool {
-    let mut index = 0;
-    while index < args.len() {
-        match args[index].as_str() {
-            "--no-pager" | "--no-optional-locks" => index += 1,
-            "-C" | "-c" | "--git-dir" | "--work-tree" => return false,
-            value if value.starts_with('-') => return false,
-            _ => break,
-        }
-    }
-    let Some(subcommand) = args.get(index).map(String::as_str) else {
-        return false;
-    };
-    let rest = &args[index + 1..];
-    match subcommand {
-        "diff" | "log" | "show" | "shortlog" | "status" | "blame" | "ls-files" | "merge-base"
-        | "rev-parse" | "rev-list" | "describe" | "cat-file" | "for-each-ref" | "grep" => true,
-        "stash" => rest.first().is_some_and(|value| value == "list"),
-        "remote" => rest.is_empty() || rest == ["-v"] || rest == ["--verbose"],
-        "config" => rest.first().is_some_and(|value| value == "--get"),
-        "reflog" => !rest
-            .iter()
-            .any(|value| matches!(value.as_str(), "expire" | "delete" | "exists")),
-        "branch" => {
-            rest.is_empty()
-                || rest.iter().all(|value| {
-                    matches!(
-                        value.as_str(),
-                        "--list" | "-l" | "-a" | "--all" | "-r" | "--remotes" | "-v" | "-vv"
-                    )
-                })
-        }
-        _ => false,
-    }
-}
-
-fn docker_args_are_read_only(args: &[String]) -> bool {
-    args.first()
-        .is_some_and(|value| matches!(value.as_str(), "ps" | "images" | "logs" | "inspect"))
-}
-
-fn split_shell_segments(command: &str) -> Option<Vec<String>> {
-    let mut segments = Vec::new();
-    let mut current = String::new();
-    let mut chars = command.chars().peekable();
-    let mut quote: Option<char> = None;
-    while let Some(ch) = chars.next() {
-        if let Some(active_quote) = quote {
-            if ch == active_quote {
-                quote = None;
-            }
-            current.push(ch);
-            continue;
-        }
-        match ch {
-            '\'' | '"' => {
-                quote = Some(ch);
-                current.push(ch);
-            }
-            ';' | '|' => {
-                push_shell_segment(&mut segments, &mut current);
-            }
-            '&' if chars.peek() == Some(&'&') => {
-                chars.next();
-                push_shell_segment(&mut segments, &mut current);
-            }
-            '&' => return None,
-            _ => current.push(ch),
-        }
-    }
-    if quote.is_some() {
-        return None;
-    }
-    push_shell_segment(&mut segments, &mut current);
-    Some(segments)
-}
-
-fn push_shell_segment(segments: &mut Vec<String>, current: &mut String) {
-    let segment = current.trim();
-    if !segment.is_empty() {
-        segments.push(segment.to_string());
-    }
-    current.clear();
-}
-
-fn tokenize_shell_segment(segment: &str) -> Option<Vec<String>> {
-    let mut tokens = Vec::new();
-    let mut current = String::new();
-    let mut chars = segment.chars().peekable();
-    let mut quote: Option<char> = None;
-    while let Some(ch) = chars.next() {
-        match quote {
-            Some(active_quote) => {
-                if ch == active_quote {
-                    quote = None;
-                } else if ch == '\\' && active_quote == '"' {
-                    if let Some(next) = chars.next() {
-                        current.push(next);
-                    }
-                } else {
-                    current.push(ch);
-                }
-            }
-            None => match ch {
-                '\'' | '"' => quote = Some(ch),
-                '\\' => {
-                    if let Some(next) = chars.next() {
-                        current.push(next);
-                    }
-                }
-                '<' => return None,
-                value if value.is_whitespace() => {
-                    if !current.is_empty() {
-                        tokens.push(std::mem::take(&mut current));
-                    }
-                }
-                _ => current.push(ch),
-            },
-        }
-    }
-    if quote.is_some() {
-        return None;
-    }
-    if !current.is_empty() {
-        tokens.push(current);
-    }
-    Some(tokens)
-}
-
-impl BackgroundTaskStore {
-    pub fn new(dir: PathBuf) -> Result<Self, ToolError> {
-        std::fs::create_dir_all(&dir)?;
-        Ok(Self {
-            dir,
-            tasks: Arc::new(Mutex::new(HashMap::new())),
-            stop_signals: Arc::new(Mutex::new(HashMap::new())),
-        })
-    }
-
-    fn start_record(
-        &self,
-        command: String,
-        sandboxed: bool,
-        sandbox_backend: String,
-    ) -> Result<(BackgroundTaskRecord, oneshot::Receiver<()>), ToolError> {
-        let id = format!("bash-{}", Uuid::new_v4());
-        let output_path = self.dir.join(format!("{id}.log"));
-        let record = BackgroundTaskRecord {
-            id: id.clone(),
-            command,
-            output_path,
-            status: BackgroundTaskStatus::Running,
-            exit_code: None,
-            sandboxed,
-            sandbox_backend,
-        };
-        let (stop_tx, stop_rx) = oneshot::channel();
-        self.tasks
-            .lock()
-            .expect("background task store lock")
-            .insert(id.clone(), record.clone());
-        self.stop_signals
-            .lock()
-            .expect("background task stop signal lock")
-            .insert(id, stop_tx);
-        Ok((record, stop_rx))
-    }
-
-    fn finish(&self, id: &str, status: BackgroundTaskStatus, exit_code: Option<i32>) {
-        if let Some(record) = self
-            .tasks
-            .lock()
-            .expect("background task store lock")
-            .get_mut(id)
-        {
-            if !matches!(record.status, BackgroundTaskStatus::Killed) {
-                record.status = status;
-            }
-            record.exit_code = exit_code;
-        }
-        self.stop_signals
-            .lock()
-            .expect("background task stop signal lock")
-            .remove(id);
-    }
-
-    fn get(&self, id: &str) -> Option<BackgroundTaskRecord> {
-        self.tasks
-            .lock()
-            .expect("background task store lock")
-            .get(id)
-            .cloned()
-    }
-
-    fn list(&self) -> Vec<BackgroundTaskRecord> {
-        let mut records = self
-            .tasks
-            .lock()
-            .expect("background task store lock")
-            .values()
-            .cloned()
-            .collect::<Vec<_>>();
-        records.sort_by(|left, right| left.id.cmp(&right.id));
-        records
-    }
-
-    fn stop(&self, id: &str) -> Result<BackgroundTaskRecord, ToolError> {
-        let mut tasks = self.tasks.lock().expect("background task store lock");
-        let record = tasks
-            .get_mut(id)
-            .ok_or_else(|| ToolError::InvalidInput(format!("unknown task id: {id}")))?;
-        if !matches!(record.status, BackgroundTaskStatus::Running) {
-            return Ok(record.clone());
-        }
-        record.status = BackgroundTaskStatus::Killed;
-        let stopped = record.clone();
-        drop(tasks);
-
-        if let Some(stop) = self
-            .stop_signals
-            .lock()
-            .expect("background task stop signal lock")
-            .remove(id)
-        {
-            let _ = stop.send(());
-        }
-        Ok(stopped)
-    }
-
-    fn stop_all(&self) -> Vec<BackgroundTaskRecord> {
-        let ids = self
-            .list()
-            .into_iter()
-            .filter(|record| matches!(record.status, BackgroundTaskStatus::Running))
-            .map(|record| record.id)
-            .collect::<Vec<_>>();
-        ids.into_iter()
-            .filter_map(|id| self.stop(&id).ok())
-            .collect()
-    }
-}
-
-fn sandbox_command_env(
-    sandbox_home: &Path,
-    base_env: &HashMap<String, String>,
-    overrides: &HashMap<String, String>,
-) -> HashMap<String, String> {
-    let sandbox_home = sandbox_home.to_string_lossy();
-    let mut env_map = HashMap::from([
-        ("HOME".to_string(), sandbox_home.to_string()),
-        (
-            "XDG_CONFIG_HOME".to_string(),
-            format!("{sandbox_home}/.config"),
-        ),
-        (
-            "XDG_CACHE_HOME".to_string(),
-            format!("{sandbox_home}/.cache"),
-        ),
-        (
-            "XDG_STATE_HOME".to_string(),
-            format!("{sandbox_home}/.local/state"),
-        ),
-        (
-            "XDG_DATA_HOME".to_string(),
-            format!("{sandbox_home}/.local/share"),
-        ),
-    ]);
-    env_map.extend(
-        base_env
-            .iter()
-            .map(|(key, value)| (key.clone(), value.clone())),
-    );
-    env_map.extend(
-        overrides
-            .iter()
-            .map(|(key, value)| (key.clone(), value.clone())),
-    );
-    ensure_usable_path(&mut env_map);
-    env_map
-}
-
-fn ensure_usable_path(env_map: &mut HashMap<String, String>) {
-    let needs_path = env_map.get("PATH").map_or(true, |value| value.is_empty());
-    if needs_path {
-        let fallback_path = env::var("PATH")
-            .ok()
-            .filter(|value| !value.is_empty())
-            .unwrap_or_else(|| "/usr/bin:/bin".to_string());
-        env_map.insert("PATH".to_string(), fallback_path);
-    }
-}
-
-fn command_env_for_wrapped(
-    wrapped: &WrappedCommand,
-    base_env: &HashMap<String, String>,
-    overrides: &HashMap<String, String>,
-) -> Result<HashMap<String, String>, ToolError> {
-    if wrapped.sandboxed {
-        let sandbox_home = wrapped.sandbox_home.as_deref().ok_or_else(|| {
-            ToolError::ExecutionFailed("sandboxed command is missing sandbox home".into())
-        })?;
-        Ok(sandbox_command_env(sandbox_home, base_env, overrides))
-    } else {
-        let mut env_map = HashMap::with_capacity(base_env.len() + overrides.len());
-        env_map.extend(
-            base_env
-                .iter()
-                .map(|(key, value)| (key.clone(), value.clone())),
-        );
-        env_map.extend(
-            overrides
-                .iter()
-                .map(|(key, value)| (key.clone(), value.clone())),
-        );
-        Ok(env_map)
-    }
 }
 
 #[async_trait]
@@ -901,323 +237,6 @@ impl Tool for BashTool {
     }
 }
 
-#[async_trait]
-impl Tool for BackgroundTaskListTool {
-    fn name(&self) -> &str {
-        "background_task_list"
-    }
-
-    fn description(&self) -> &str {
-        "List background bash tasks"
-    }
-
-    fn input_schema(&self) -> Value {
-        json!({
-            "type": "object",
-            "properties": {}
-        })
-    }
-
-    async fn call(&self, _input: Value) -> Result<Value, ToolError> {
-        Ok(json!({
-            "tasks": self.background_tasks.list(),
-        }))
-    }
-}
-
-#[async_trait]
-impl Tool for BackgroundTaskStatusTool {
-    fn name(&self) -> &str {
-        "background_task_status"
-    }
-
-    fn description(&self) -> &str {
-        "Inspect a background bash task and read the tail of its output"
-    }
-
-    fn input_schema(&self) -> Value {
-        json!({
-            "type": "object",
-            "properties": {
-                "task_id": {
-                    "type": "string",
-                    "description": "Background task id returned by bash run_in_background."
-                },
-                "tail_bytes": {
-                    "type": "integer",
-                    "minimum": 1,
-                    "default": 12000,
-                    "description": "Maximum number of output bytes to return from the end of the task log."
-                }
-            },
-            "required": ["task_id"]
-        })
-    }
-
-    async fn call(&self, input: Value) -> Result<Value, ToolError> {
-        let task_id = input["task_id"]
-            .as_str()
-            .ok_or_else(|| ToolError::InvalidInput("task_id".into()))?;
-        let tail_bytes = input
-            .get("tail_bytes")
-            .and_then(Value::as_u64)
-            .unwrap_or(12_000)
-            .min(1_000_000) as usize;
-        let record = self
-            .background_tasks
-            .get(task_id)
-            .ok_or_else(|| ToolError::InvalidInput(format!("unknown task id: {task_id}")))?;
-        let output = read_output_tail(&record.output_path, tail_bytes).await?;
-
-        Ok(json!({
-            "task_id": record.id,
-            "command": record.command,
-            "status": record.status,
-            "exit_code": record.exit_code,
-            "output_path": record.output_path,
-            "output": output,
-            "sandboxed": record.sandboxed,
-            "sandbox_backend": record.sandbox_backend,
-        }))
-    }
-}
-
-#[async_trait]
-impl Tool for BackgroundTaskStopTool {
-    fn name(&self) -> &str {
-        "background_task_stop"
-    }
-
-    fn description(&self) -> &str {
-        "Stop one background bash task, or all running background bash tasks when task_id is omitted"
-    }
-
-    fn input_schema(&self) -> Value {
-        json!({
-            "type": "object",
-            "properties": {
-                "task_id": {
-                    "type": "string",
-                    "description": "Background task id returned by bash run_in_background. Omit to stop all running background bash tasks."
-                }
-            }
-        })
-    }
-
-    async fn call(&self, input: Value) -> Result<Value, ToolError> {
-        if let Some(task_id) = input.get("task_id").and_then(Value::as_str) {
-            let task = self.background_tasks.stop(task_id)?;
-            return Ok(json!({ "stopped": [task] }));
-        }
-        Ok(json!({ "stopped": self.background_tasks.stop_all() }))
-    }
-}
-
-fn spawn_background_bash_task(
-    mut child: Child,
-    wrapped: WrappedCommand,
-    record: BackgroundTaskRecord,
-    store: Arc<BackgroundTaskStore>,
-    stop_rx: oneshot::Receiver<()>,
-) {
-    tokio::spawn(async move {
-        let result = run_background_bash_task(&mut child, wrapped, &record, stop_rx).await;
-        let (status, exit_code) = match result {
-            Ok(code) => {
-                if code == Some(0) {
-                    (BackgroundTaskStatus::Completed, code)
-                } else {
-                    (BackgroundTaskStatus::Failed, code)
-                }
-            }
-            Err(err) => {
-                let _ = append_background_output(
-                    &record.output_path,
-                    BashStreamKind::Stderr,
-                    &format!("background task failed: {err}\n"),
-                )
-                .await;
-                (BackgroundTaskStatus::Failed, None)
-            }
-        };
-        store.finish(&record.id, status, exit_code);
-    });
-}
-
-async fn run_background_bash_task(
-    child: &mut Child,
-    wrapped: WrappedCommand,
-    record: &BackgroundTaskRecord,
-    mut stop_rx: oneshot::Receiver<()>,
-) -> Result<Option<i32>, ToolError> {
-    if let Some(parent) = record.output_path.parent() {
-        fs::create_dir_all(parent).await?;
-    }
-    fs::write(&record.output_path, "").await?;
-    if !wrapped.sandboxed {
-        append_background_output(
-            &record.output_path,
-            BashStreamKind::Stderr,
-            &unsandboxed_execution_warning(&wrapped),
-        )
-        .await?;
-    }
-
-    let stdout = child
-        .stdout
-        .take()
-        .ok_or_else(|| ToolError::ExecutionFailed("stdout pipe unavailable".into()))?;
-    let stderr = child
-        .stderr
-        .take()
-        .ok_or_else(|| ToolError::ExecutionFailed("stderr pipe unavailable".into()))?;
-    let (tx, mut rx) = mpsc::channel(64);
-    let stdout_task = tokio::spawn(read_stream_chunks(
-        stdout,
-        BashStreamKind::Stdout,
-        tx.clone(),
-    ));
-    let stderr_task = tokio::spawn(read_stream_chunks(stderr, BashStreamKind::Stderr, tx));
-
-    let mut output_file = fs::OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(&record.output_path)
-        .await?;
-    let mut stop_requested = false;
-    loop {
-        tokio::select! {
-            chunk = rx.recv() => {
-                let Some((stream, chunk)) = chunk else {
-                    break;
-                };
-                if !chunk.is_empty() {
-                    match stream {
-                        BashStreamKind::Stdout => output_file.write_all(chunk.as_bytes()).await?,
-                        BashStreamKind::Stderr => {
-                            output_file.write_all(b"[stderr] ").await?;
-                            output_file.write_all(chunk.as_bytes()).await?;
-                        }
-                    }
-                }
-            }
-            _ = &mut stop_rx, if !stop_requested => {
-                stop_requested = true;
-                child.start_kill()
-                    .map_err(|err| ToolError::ExecutionFailed(format!("stop background task: {err}")))?;
-                output_file.write_all(b"[stderr] background task stop requested\n").await?;
-            }
-        }
-    }
-
-    let status = child.wait().await?;
-    stdout_task
-        .await
-        .map_err(|err| ToolError::ExecutionFailed(err.to_string()))??;
-    stderr_task
-        .await
-        .map_err(|err| ToolError::ExecutionFailed(err.to_string()))??;
-    if let Some(path) = wrapped.cleanup_path.as_ref() {
-        let _ = fs::remove_file(path).await;
-    }
-    Ok(status.code())
-}
-
-async fn append_background_output(
-    path: &Path,
-    stream: BashStreamKind,
-    chunk: &str,
-) -> Result<(), ToolError> {
-    let mut file = fs::OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(path)
-        .await?;
-    match stream {
-        BashStreamKind::Stdout => file.write_all(chunk.as_bytes()).await?,
-        BashStreamKind::Stderr => {
-            file.write_all(b"[stderr] ").await?;
-            file.write_all(chunk.as_bytes()).await?;
-        }
-    }
-    Ok(())
-}
-
-async fn read_output_tail(path: &Path, max_bytes: usize) -> Result<String, ToolError> {
-    let mut file = match fs::File::open(path).await {
-        Ok(file) => file,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(String::new()),
-        Err(err) => return Err(err.into()),
-    };
-    let file_len = file.metadata().await?.len();
-    let start = file_len.saturating_sub(max_bytes as u64);
-    file.seek(SeekFrom::Start(start)).await?;
-    let mut bytes = Vec::with_capacity(max_bytes.min(file_len as usize));
-    file.read_to_end(&mut bytes).await?;
-    Ok(String::from_utf8_lossy(&bytes).into_owned())
-}
-
-async fn read_stream_chunks<R>(
-    reader: R,
-    stream: BashStreamKind,
-    tx: mpsc::Sender<(BashStreamKind, String)>,
-) -> Result<(), ToolError>
-where
-    R: AsyncRead + Unpin + Send + 'static,
-{
-    let mut reader = reader;
-    let mut buffer = [0_u8; 4096];
-    loop {
-        let read = reader.read(&mut buffer).await?;
-        if read == 0 {
-            break;
-        }
-        let chunk = String::from_utf8_lossy(&buffer[..read]).into_owned();
-        if tx.send((stream, chunk)).await.is_err() {
-            break;
-        }
-    }
-    Ok(())
-}
-
-fn sandbox_output_hint(stderr: &str) -> Option<&'static str> {
-    let lower = stderr.to_ascii_lowercase();
-    if lower.contains("sandbox: violation")
-        || lower.contains("operation not permitted")
-        || lower.contains("command not found")
-        || lower.contains("no such file or directory")
-        || lower.contains("permission denied")
-    {
-        Some(
-            "\n\nhint: Sandboxed bash appears blocked or missing a runtime path. Prefer direct file tools such as read_file, apply_patch, and replace_lines; ask the user only if a real shell command is required.\n",
-        )
-    } else {
-        None
-    }
-}
-
-fn unsandboxed_execution_warning(wrapped: &WrappedCommand) -> String {
-    format!(
-        "warning: command is running without sandbox isolation (backend: {}).\n",
-        wrapped.sandbox_backend
-    )
-}
-
-async fn ensure_sandbox_home_dirs(sandbox_home: &Path) -> Result<(), ToolError> {
-    for dir in [
-        sandbox_home.to_path_buf(),
-        sandbox_home.join(".config"),
-        sandbox_home.join(".cache"),
-        sandbox_home.join(".local"),
-        sandbox_home.join(".local/state"),
-        sandbox_home.join(".local/share"),
-    ] {
-        fs::create_dir_all(dir).await?;
-    }
-    Ok(())
-}
-
-#[cfg(test)]
 mod tests {
     use super::{
         BackgroundTaskListTool, BackgroundTaskStatus, BackgroundTaskStatusTool,
